Hint that onChangeText will pass "Search Text".
I just guessed that this was the case and sure enough it was, so that's pretty good design by you. :) 

However, I think it would be good to lead the reader a little bit more.

I know there is an example directory but this doesn't hurt to get people started.